% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Risk.R
\name{data}
\alias{data}
\title{A illustration dataset}
\description{
Generate a dataset, SP500 returns from https://finance.yahoo.com/, GDP growth, Inflation, Unemployment rate of China from https://data.worldbank.org.cn.
}
\examples{
\dontrun{
data(data)
attach(data)
plot(Year, SP_Return)
plot(Year, GDP)
plot(Year, Inflation)
plot(Year, Unemployment)
}
}
