% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Risk.R
\name{dataprediction}
\alias{dataprediction}
\title{Stock risk of data}
\description{
Using yearly sp500 returns to predict the stock risk value in China, by using data.
}
\examples{
\dontrun{
data(data)
attach(data)
r1 = predict_stock_market_risk(SP_Return,GDP,Inflation,Unemployment,window_size=10)$risk_predictions
plot(seq(2024,2038),r1,ylab = 'Stock Risk Value',type = 'l')
}
}
