% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Risk.R
\name{df}
\alias{df}
\title{An illustration dataset with 2 columns: pressure and temp to be detected}
\description{
A simple example dataset, detect the outliers of first column temp and the 2nd column pressure.
}
\examples{
\dontrun{
data(df)
attach(df)
plot(temp)
plot(pressure)
}
}
