% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Risk.R
\name{predict_stock_market_risk}
\alias{predict_stock_market_risk}
\title{Forecast stock market risk based on macroeconomic data}
\usage{
predict_stock_market_risk(stock_returns, x1, x2, x3, window_size = 80)
}
\arguments{
\item{stock_returns}{stock returns}

\item{x1}{The first factor, such as GDP growth}

\item{x2}{The second factor, such as inflation rate}

\item{x3}{The third factor, such as umemployment rate}

\item{window_size}{The size of the scroll window, 50 by default}
}
\value{
prediction of risk value of size n-window_size+1.
}
\description{
Combine gdp, unemployment rate and inflation rate to predict future stock market risk
}
\examples{
\dontrun{
stock_returns <- rnorm(200)
gdp_growth <- rnorm(200)
inflation_rate <- rnorm(200)
unemployment_rate <- rnorm(200) 
predict_stock_market_risk(stock_returns, gdp_growth, inflation_rate, unemployment_rate)
}
}
