% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rangeAnomalyDetection}
\alias{rangeAnomalyDetection}
\title{rangeAnomalyDetection}
\usage{
rangeAnomalyDetection(df, value_ranges = NULL)
}
\arguments{
\item{df}{data to be detected if there are outliers, type of data frame}

\item{value_ranges}{data out of this normal range will be extracted, IQR(interquantile range) by default.}
}
\value{
a list, including the index of anomaly, and the range.
}
\description{
Outlier detection by range
}
\examples{
\dontrun{
df2 <- data.frame(
d1 = runif(10, -30, 40),
d2 = runif(10, 900, 1100)
)
rangeAnomalyDetection(df2, value_ranges)
}
}
