\name{SA24204124-package}
\alias{SA24204124-package}
\docType{package}
\title{
  Two functions and homework answers
}
\description{
  Several functions are included in this package for illustration.
  }
\section{Package Content}{\packageIndices{SA24204124}}
\author{
Yuning Dong
Maintainer: Yuning Dong <sa24204124@mail.ustc.edu.cn>
}
\section{Maintainer}{\packageMaintainer{SA24204124}}
\keyword{ package }
